% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_xts}
\alias{aggregate_xts}
\title{Aggregate data to the given time granularity}
\usage{
aggregate_xts(data, on, k = 1, FUN = "sum", na_handling = "auto",
  align_data = TRUE, start_date = NULL, end_date = NULL, ...)
}
\arguments{
\item{data}{Input \code{\link{xts}} object.}

\item{on}{Determines the new unit of the data frequency. Can be "minutes", "hours", or "days".}

\item{k}{The new value of frequency (only for hours and minutes).}

\item{FUN}{Function to aggregate the data with. Either a closure (function), or a string.
Strings currently supported are "sum", "prod", "min", "max", last", and "first".
By default \code{sum} is used.}

\item{na_handling}{What to do with NA's. "auto" fills 0 if FUN == sum, na.locf's last value else.
NULL does nothin to NAs. If a function is given, it is applied directly.}

\item{align_data}{Align data to the next truncated time-stamp}

\item{start_date}{First Date in the regular series. Calculated from the given series if NULL. Its type must match the type of \code{data}'s index.}

\item{end_date}{Last Date in the regular series. Calculated from the given series if NULL. Its type must match the type of \code{data}'s index.}

\item{...}{Arguements passed to \code{FUN}.}
}
\value{
Aggregated Time-series
}
\description{
Aggregate data to the given time granularity
}

