% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{binary_hour_matrix}
\alias{binary_hour_matrix}
\title{Create dummy variables for hour of day}
\usage{
binary_hour_matrix(indices, no_dependency = TRUE)
}
\arguments{
\item{indices}{Input time-stamps}

\item{no_dependency}{Remove linear dependency between variables}
}
\value{
An xts object
}
\description{
Create dummy variables for hour of day
}

