% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{build_mltsp_pipline}
\alias{build_mltsp_pipline}
\title{Create an ML TSP pipeline object}
\usage{
build_mltsp_pipline(preproc = NULL, featext, learner, predictor = predict)
}
\arguments{
\item{preproc}{list of preprocessing technique tuples}

\item{featext}{a function to extract features from a time-series}

\item{learner}{a function to create the ML model}

\item{predictor}{ML prediction function (\code{\link{predict}} by default)}
}
\value{
An MLTS pipeline object
}
\description{
Create an ML TSP pipeline object
}
\seealso{
\code{\link{mltsp}}, \code{\link{mltsp_forecaster}}
}

