% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_window.R
\name{centered_lagged_windows}
\alias{centered_lagged_windows}
\title{Create Centers the lag winow around Lag 0, with width p on each side.}
\usage{
centered_lagged_windows(x, p = 1, pstr = deparse(substitute(x)))
}
\arguments{
\item{x}{data input (can handle multi-column data, such as data.frames)}

\item{p}{window length around lag 0.}

\item{pstr}{name of data column}
}
\value{
A time-series object of lagged and leaded x. If \code{p == 0}, \code{x} itself is returned. If \code{p < 0}, \code{NULL} is returned.
}
\description{
Create Centers the lag winow around Lag 0, with width p on each side.
}
\details{
Use this function for creating features from time-series that are available in future.
}
\seealso{
\code{\link{lag_windows}}, \code{\link{seasonal_lag_windows}}
}

