% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_tools.R
\name{clone_timestamps}
\alias{clone_timestamps}
\title{Change values in a time-series}
\usage{
clone_timestamps(x, newdata)
}
\arguments{
\item{x}{Time-series to clone stamps from}

\item{newdata}{The new values}
}
\description{
Change values in a time-series
}

