% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_tools.R
\name{continue_timestamps}
\alias{continue_timestamps}
\title{Continue the given time-stamps for the next \code{h} steps}
\usage{
continue_timestamps(timestamps, h = 1)
}
\arguments{
\item{timestamps}{The timestamps to continue}

\item{h}{Horizon len to forecast the new time-series}
}
\value{
A vector of \code{h} time-stamps
}
\description{
Continue the given time-stamps for the next \code{h} steps
}

