% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_tools.R
\name{continue_timestamps_using_reference}
\alias{continue_timestamps_using_reference}
\title{Continue the given time-stamps for the next \code{h} steps using a reference timestamp}
\usage{
continue_timestamps_using_reference(timestamps, h = 1,
  reference.stamps = NULL)
}
\arguments{
\item{timestamps}{The timestamps to continue}

\item{h}{Horizon len to forecast the new time-series}

\item{reference.stamps}{Reference time-stamps. If given, the first of these time-stamps which are
larger than the last of "timestamps" are chosen and used}
}
\value{
A vector of \code{h} time-stamps
}
\description{
Continue the given time-stamps for the next \code{h} steps using a reference timestamp
}

