% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_time_slices.R
\name{create_cv_timeslices}
\alias{create_cv_timeslices}
\title{Create time-slices for time-series cross-validation}
\usage{
create_cv_timeslices(y, initial_window, horizon = 1, fixed_window = TRUE,
  skip = 0)
}
\arguments{
\item{y}{The input time-series}

\item{initial_window}{initial number of consecutive values in each training set sample}

\item{horizon}{number of consecutive values in test set sample}

\item{fixed_window}{To use a fixed window of size \code{initial_window} for training every stride. If FALSE, we use the maximum possible length for the training set.}

\item{skip}{How many slices are to be skipped and not used for cross-validation}
}
\value{
Time-slices, for use with \code{\link{ts_crossval}}
}
\description{
Create time-slices for time-series cross-validation
}
\details{
It is ensured that \code{initial_windowlength + horizonlength <= length(y)}.
}
\references{
See \url{https://github.com/topepo/caret/blob/master/pkg/caret/R/createTimeSlices.R}
}

