% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{create_hourly_xts}
\alias{create_hourly_xts}
\title{Aggregate data to a hourly time-series}
\usage{
create_hourly_xts(data, hours = 1, FUN = sum, start_date = NULL,
  end_date = NULL, ...)
}
\arguments{
\item{data}{Input \code{\link{xts}} object.}

\item{hours}{Hours seperating two samples.}

\item{FUN}{Function to aggregate the data with. By default \code{sum} is used.}

\item{start_date}{First Date in the regular series. Calculated from the given series if not given.}

\item{end_date}{Last Date in the regular series. Calculated from the given series if not given.}

\item{...}{Other parameters given to FUN}
}
\value{
Hourly data time Series
}
\description{
Aggregate data to a hourly time-series
}

