% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{forecast.mltsp}
\alias{forecast.mltsp}
\title{Machine learning pipeline forecast}
\usage{
\method{forecast}{mltsp}(object, h, test_timestamps = NULL, ...)
}
\arguments{
\item{object}{A mlts model}

\item{h}{Horizon of forecast}

\item{test_timestamps}{Time-stamps of forecast. Will be generated automatically if \code{NULL}.}

\item{...}{Extra parameters passed to feature extraction, prediction, and inverse pre-processing in the pipeline}
}
\description{
Takes a \code{\link{mltsp}} object as an input. The machine learning should be configured in this object.
}

