% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narx.R
\name{forecast.narx}
\alias{forecast.narx}
\title{Forecast a NARX object}
\usage{
\method{forecast}{narx}(object, h = 10, level = c(80, 95), fan = FALSE,
  xreg = NULL, lambda = object$lambda, ...)
}
\arguments{
\item{object}{An object of class "narx". Usually the result of a call to \code{\link{narx}}.}

\item{h}{Number of periods for forecasting. If \code{xreg} is used, \code{h} is ignored and the number of forecast periods is set to the number of rows of \code{xreg}.}

\item{level}{Not used}

\item{fan}{Not used}

\item{xreg}{Future values of an exogenous regression variables.}

\item{lambda}{Box-Cox transformation parameter. Ignored if \code{NULL}. Otherwise, forecasts back-transformed via an inverse Box-Cox transformation.}

\item{...}{Inputs to the learning function's \code{\link{predict}}}
}
\value{
A \code{\link{forecast}} package compatible \code{forecast} object
}
\description{
Forecast a NARX object
}

