% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_window.R
\name{lagmatrix}
\alias{lagmatrix}
\title{Collects and embeds lagged windows of data.}
\usage{
lagmatrix(x, lag_indices, pstr = deparse(substitute(x)), no_peeking = FALSE,
  ...)
}
\arguments{
\item{x}{data input (can handle multi-column data, such as data.frames)}

\item{lag_indices}{index of lags to embed. Can be empty}

\item{pstr}{name of datacolumn (only used if x is a single column or vector)}

\item{no_peeking}{if TRUE, lags <= 0 (from now and future) are removed}

\item{...}{Not used}
}
\value{
a matrix of lagged windows
}
\description{
Collects and embeds lagged windows of data.
}

