% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{mltsp}
\alias{mltsp}
\alias{mltsp.default}
\alias{mltsp.mltsp}
\alias{mltsp.mltsp_pipeline}
\title{Create an ML TSP object}
\usage{
mltsp(...)

\method{mltsp}{mltsp}(object, x, ...)

\method{mltsp}{default}(x, preproc = NULL, featext, learner,
  predictor = predict, ...)

\method{mltsp}{mltsp_pipeline}(pipeline, x, ...)
}
\arguments{
\item{...}{Other parameters passed to feature extraction, learning, and prediction function}

\item{object}{A mltsp object}

\item{x}{xts time-series for training}

\item{preproc}{list of preprocessing technique tuples}

\item{featext}{a function to extract features from a time-series}

\item{learner}{a function to create the ML model}

\item{predictor}{ML prediction function (\code{\link{predict}} by default)}

\item{pipeline}{A ML TSP pipeline object}
}
\value{
A fitted ML TSP model
}
\description{
Create an ML TSP object
}
\seealso{
\code{\link{build_mltsp_pipline}}, \code{\link{mltsp_forecaster}}
}

