% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{mltsp_forecaster}
\alias{mltsp_forecaster}
\title{Functional version of the mlts}
\usage{
mltsp_forecaster(preproc = NULL, featext, learner, predictor = predict, ...)
}
\arguments{
\item{preproc}{list of preprocessing technique tuples}

\item{featext}{a function to extract features from a time-series}

\item{learner}{a function to create the ML model}

\item{predictor}{forecast function (\code{\link{predict}} by default)}

\item{...}{Other parameters passed to \code{\link{build_mltsp_pipline}}}
}
\value{
A \code{function(x, h, ...)} that takes a time-series \code{x} and a horizon \code{h},
and returns the forecast as another time-series.
}
\description{
Functional version of the mlts
}
\seealso{
\code{\link{build_mltsp_pipline}}, \code{\link{mltsp}}
}

