% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narx.R
\name{narx}
\alias{narx}
\alias{narx.default}
\alias{narx.narx}
\title{Nonlinear auto-regressive time-series forecasting with exogenous input}
\usage{
narx(...)

\method{narx}{narx}(model, x, xreg = NULL, ...)

\method{narx}{default}(x, learner_func, p, P = 0, d = 0, D = 0,
  freq = 1, xreg = NULL, lambda = NULL, ...)
}
\arguments{
\item{...}{Extra parameters passed to \code{learner_func}}

\item{model}{NARX model. A \code{narx} object.}

\item{x}{Time-series input}

\item{xreg}{External regression data}

\item{learner_func}{Non-linear learning function}

\item{p}{Auto-regressive order}

\item{P}{Seasonal auto-regressive order}

\item{d}{Model differencing order}

\item{D}{Seasonal differencing order}

\item{freq}{Time-series periodicity}

\item{lambda}{Box-Cox transform's lambda}
}
\value{
A fitted NARX object
}
\description{
Nonlinear auto-regressive time-series forecasting with exogenous input
}
\seealso{
\code{\link{build_narx}}
}

