% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_tools.R
\name{period.apply.string}
\alias{period.apply.string}
\title{Period apply with string function}
\usage{
period.apply.string(x, INDEX, FUN)
}
\arguments{
\item{x}{a univariate data object}

\item{INDEX}{a numeric vector of endpoints to calculate maximum on}

\item{FUN}{A string. Currently supported are "sum", "prod", "min", "max", last", and "first".}
}
\value{
An xts or zoo object of maximums, indexed by the period endpoints.
}
\description{
xts's period.apply is optimized for a few function. The current function takes these as
an arguement, and then calls the correct function.
}

