% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_tools.R
\name{period.last}
\alias{period.last}
\title{Period apply with string function}
\usage{
period.last(x, INDEX)
}
\arguments{
\item{x}{a univariate data object}

\item{INDEX}{a numeric vector of endpoints to calculate maximum on}
}
\value{
An xts or zoo object of maximums, indexed by the period endpoints.
}
\description{
Period.apply with last function
}

