% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narx.R
\name{predict.narx}
\alias{predict.narx}
\title{Predict a time-series using a NARX object}
\usage{
\method{predict}{narx}(object, x, h = 10, xreg = NULL,
  lambda = object$lambda, ...)
}
\arguments{
\item{object}{An object of class "narx". Usually the result of a call to \code{\link{narx}}.}

\item{x}{A time-series to forecast}

\item{h}{Number of periods for forecasting. If \code{xreg} is used, \code{h} is ignored and the number of forecast periods is set to the number of rows of \code{xreg}.}

\item{xreg}{Future values of an exogenous regression variables. Overrides \code{h} and its timestamps are used for prediction time-stamps.}

\item{lambda}{Box-Cox transformation parameter. Ignored if \code{NULL}. Otherwise, forecasts back-transformed via an inverse Box-Cox transformation.}

\item{...}{Inputs to the learning function's \code{\link{predict}}}
}
\value{
The predicted time-series
}
\description{
Predict a time-series using a NARX object
}

