% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_window.R
\name{seasonal_lag_windows}
\alias{seasonal_lag_windows}
\title{Creates a window around each seasonal lag, with \code{width+shift}/\code{width-shift} on each side.}
\usage{
seasonal_lag_windows(x, P = 1, freq = frequency(x), width = 0,
  shift = 0, no_peeking = TRUE, pstr = deparse(substitute(x)))
}
\arguments{
\item{x}{data input (can handle multi-column data, such as data.frames)}

\item{P}{seasonal window length}

\item{freq}{seasonal frequency}

\item{width}{width of the left or right window}

\item{shift}{shifts each window towards the end of time-series}

\item{no_peeking}{if TRUE, lags <= 0 (from now and future) are removed}

\item{pstr}{name of data column}
}
\value{
A time-series object of lagged and leaded x
}
\description{
Creates a window around each seasonal lag, with \code{width+shift}/\code{width-shift} on each side.
}
\seealso{
\code{\link{lag_windows}}, \code{\link{centered_lagged_windows}}
}

