% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleLM.R
\name{SimpleLM}
\alias{SimpleLM}
\alias{predict.SimpleLM}
\title{A Simple linear model.}
\usage{
SimpleLM(X, y = NULL, ...)

\method{predict}{SimpleLM}(object, newdata, ...)
}
\arguments{
\item{X}{Input data for training}

\item{y}{Target (output) data for training. If \code{NULL}, first column of X is used.}

\item{...}{Other parameters passed to \code{\link{lm}}}

\item{object}{A \code{SimpleLM} object}

\item{newdata}{Input data for prediction}
}
\value{
a \code{\link{lm}} model
}
\description{
Replaces lm where X and y are separate
}

