% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_tools.R
\name{ts_append}
\alias{ts_append}
\title{Appends new values to a time-series.}
\usage{
ts_append(timeseries, newdata, new_timestamps = NULL)
}
\arguments{
\item{timeseries}{the timeseries to continue its timestamps}

\item{newdata}{new values to append to the timeseries}

\item{new_timestamps}{new timestamp of the data to use}
}
\value{
Time-series (\code{ts} or \code{xts}) with the new data appended to it.
}
\description{
Appends new values to a time-series. Timestamps of the new data can be given.
If not not, they are automatically deduced by from the given series data.
}

