% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_tools.R
\name{ts_continue}
\alias{ts_continue}
\title{Create a bew time-series with the new value that continues the input time-series}
\usage{
ts_continue(timeseries, newdata)
}
\arguments{
\item{timeseries}{the timeseries to continue its timestamps}

\item{newdata}{new values to append to the timeseries}
}
\value{
Time-series with the new data
}
\description{
Create a bew time-series with the new value that continues the input time-series
}

