% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{ts_crossval_simple}
\alias{ts_crossval_simple}
\title{Simple time-series cross-validation}
\usage{
ts_crossval_simple(y, func, ..., slices, error_func = rmse, break_err = Inf,
  verbose = FALSE)
}
\arguments{
\item{y}{The time-series}

\item{func}{Forecasting function. See details.}

\item{...}{Forcasting function}

\item{slices}{Cross-validation slices, created by \code{\link{create_cv_timeslices}}}

\item{error_func}{Error function for reporting cross-validation error.}

\item{break_err}{If CV error passes this value, cross-validation is prematurely stopped.}

\item{verbose}{If True, verbose debugging messages are printed out.}
}
\value{
Cross-validation error for the given forecast function
}
\description{
Simple time-series cross-validation
}
\details{
\code{func} expects a signature of type \code{function(y, h, ...)}, where \code{y} is the time-series to forecast,
\code{h} is the horizon to forecast, and \code{...} is any other parameter.

Use \code{break_err} when you have to evaluate many functions quickly.
By setting this value to CV from a known good predictor, bad forecasters can be pruned quickly.
}

