% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ts_inv_preprocess}
\alias{ts_inv_preprocess}
\title{Inverse preprocessing of  a time-series}
\usage{
ts_inv_preprocess(x, technique_list)
}
\arguments{
\item{x}{Preprocessed time-series}

\item{technique_list}{List of the techniques. Usually taken from output of \code{\link{ts_preprocess}}}
}
\value{
An xts object
}
\description{
Inverse preprocessing of  a time-series
}

