% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ts_preprocess}
\alias{ts_preprocess}
\title{Preprocess a time-series}
\usage{
ts_preprocess(x, technique_list = list())
}
\arguments{
\item{x}{Input time-series}

\item{technique_list}{List of the techniques. See details.}
}
\value{
A list
\itemize{
 \item{\code{x}}{Preprocessed xts}
 \item{\code{technique_list}}{List of techniques with their explicit parameters}
}
}
\description{
Preprocess a time-series
}
\details{
Each technique is itself a list of \code{list("techninque name", param1, param2, ...)}.
Available techniques include: \itemize{
 \item{"diff"}{Differencing the time-series. Parameter can be a number, "kpss", "adf", or "auto".}
 \item{"boxcox"}{Box-cox transform. Parameter can be a numeric lambda, or "auto".}
 \item{"log"}{Log transform. Only works if time-series is positive.}
 \item{"log_abs"}{Log transform on absolute value. Destroys values between -1 and 1.}
 \item{"log+min"}{Log transform on (x - min(x) + 1)}
}
}

